Define common types
