+++ packages/ui/src/common/types/accounts.ts
@@ -1,0 +1,6 @@
+export type Address = string
+
+export interface Account {
+  name: string | undefined
+  address: Address
+}
