+++ packages/ui/src/common/types/memberships.ts
@@ -1,0 +1,10 @@
+import { Account } from './index'
+
+export interface Member {
+  name: string
+  rootAccount: Account
+  controllerAccount: Account
+  handle: string
+  avatarURI?: string
+  about?: string
+}
