+++ packages/ui/src/components/selects/SelectAccount/OptionListAccount.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import { BorderRad, Shadows, Sizes, Transitions } from '../../../constants'
-import { Account } from '../../../hooks/types'
+import { Account } from '../../../common/types'
 import { OptionAccount } from './OptionAccount'
 
 interface Props {
