+++ packages/ui/src/hooks/useSignAndSendTransaction.ts
@@ -4,7 +4,7 @@ import { ISubmittableResult } from '@polkadot/types/types'
 import BN from 'bn.js'
 import { useEffect, useState } from 'react'
 import { Observable } from 'rxjs'
-import { Account } from './types'
+import { Account } from '../common/types'
 import { useApi } from './useApi'
 import { useKeyring } from './useKeyring'
 import { useObservable } from './useObservable'
