+++ packages/ui/src/modals/AddMembershipModal/SignCreateMemberModal.tsx
@@ -3,6 +3,7 @@ import { ISubmittableResult } from '@polkadot/types/types'
 import BN from 'bn.js'
 import React, { useState } from 'react'
 import { Observable } from 'rxjs'
+import { Member } from '../../common/types'
 import { ButtonPrimaryMedium } from '../../components/buttons'
 import { Label } from '../../components/forms'
 import { Help } from '../../components/Help'
@@ -12,7 +13,6 @@ import { Text, TokenValue } from '../../components/typography'
 import { useApi } from '../../hooks/useApi'
 import { useSignAndSendTransaction } from '../../hooks/useSignAndSendTransaction'
 import { BalanceInfoNarrow, InfoTitle, InfoValue, Row } from '../common'
-import { Member } from './MembershipFormModal'
 
 interface SignProps {
   onClose: () => void
