+++ packages/ui/src/modals/TransferModal/SignTransferModal.tsx
@@ -8,7 +8,7 @@ import { Help } from '../../components/Help'
 import { ArrowDownExpandedIcon } from '../../components/icons'
 import { Modal, ModalBody, ModalFooter, ModalHeader, SignTransferContainer } from '../../components/Modal'
 import { TokenValue } from '../../components/typography'
-import { Account } from '../../hooks/types'
+import { Account } from '../../common/types'
 import { useApi } from '../../hooks/useApi'
 import { useBalance } from '../../hooks/useBalance'
 import { useSignAndSendTransaction } from '../../hooks/useSignAndSendTransaction'
