+++ packages/ui/src/modals/TransferModal/TransactionSuccessModal.tsx
@@ -4,7 +4,7 @@ import { AccountInfo } from '../../components/AccountInfo'
 import { ArrowDownExpandedIcon, SuccessIcon } from '../../components/icons'
 import { ModalHeader, Modal, SuccessModalBody } from '../../components/Modal'
 import { TokenValue } from '../../components/typography'
-import { Account } from '../../hooks/types'
+import { Account } from '../../common/types'
 import { useBalance } from '../../hooks/useBalance'
 import {
   BalanceInfo,
