+++ packages/ui/src/modals/TransferModal/TransferModal.tsx
@@ -2,7 +2,7 @@ import { ISubmittableResult } from '@polkadot/types/types'
 import BN from 'bn.js'
 import React, { ReactElement, useEffect, useState } from 'react'
 import { Observable, Subscription } from 'rxjs'
-import { Account } from '../../hooks/types'
+import { Account } from '../../common/types'
 import { useKeyring } from '../../hooks/useKeyring'
 import { WaitModal } from '../WaitModal'
 import { SignTransferModal } from './SignTransferModal'
