+++ packages/ui/src/pages/Profile/MyAccounts/Accounts.tsx
@@ -6,7 +6,7 @@ import { PageTab, PageTabsNav } from '../../../components/page/PageTabs'
 import { TransferButton } from '../../../components/TransferButton'
 import { TokenValue } from '../../../components/typography'
 import { BorderRad, Colors, Sizes } from '../../../constants'
-import { Account } from '../../../hooks/types'
+import { Account } from '../../../common/types'
 import { useAccounts } from '../../../hooks/useAccounts'
 import { useBalance } from '../../../hooks/useBalance'
 import { useBalances } from '../../../hooks/useBalances'
