+++ packages/ui/test/modals/AddMembershipModal.test.tsx
@@ -8,7 +8,7 @@ import { set } from 'lodash'
 import React from 'react'
 import { from, of } from 'rxjs'
 import sinon from 'sinon'
-import { Account } from '../../src/hooks/types'
+import { Account } from '../../src/common/types'
 import * as useAccountsModule from '../../src/hooks/useAccounts'
 import { AddMembershipModal } from '../../src/modals/AddMembershipModal'
 import { ApiContext } from '../../src/providers/api/context'
