+++ packages/ui/test/pages/Accounts.test.tsx
@@ -11,7 +11,7 @@ import { KeyringContext } from '../../src/providers/keyring/context'
 import * as useAccountsModule from '../../src/hooks/useAccounts'
 import * as useBalanceModule from '../../src/hooks/useBalance'
 import { Accounts } from '../../src/pages/Profile/MyAccounts/Accounts'
-import { Account } from '../../src/hooks/types'
+import { Account } from '../../src/common/types'
 
 describe('UI: Accounts list', () => {
   let accounts: {
